+++ src/chains/Arbitrum/Common.sol
@@ -42,7 +42,7 @@ import {IAlgebraCallback} from "../../core/univ3forks/Algebra.sol";
 import {camelotV3Factory, camelotV3InitHash, camelotV3ForkId} from "../../core/univ3forks/CamelotV3.sol";
 import {dackieSwapV3ArbitrumFactory, dackieSwapV3ForkId} from "../../core/univ3forks/DackieSwapV3.sol";
 
-import {ARBITRUM_POOL_MANAGER} from "../../core/UniswapV4Addresses.sol";
+import {uniswapV4ArbitrumPoolManager, uniswapV4ForkId} from "../../core/univ4forks/UniswapV4.sol";
 
 // Solidity inheritance is stupid
 import {SettlerSwapAbstract} from "../../SettlerAbstract.sol";
@@ -82,14 +82,15 @@ abstract contract ArbitrumMixin is
                 address recipient,
                 IERC20 sellToken,
                 uint256 ppm,
+                uint8 forkId,
                 bool feeOnTransfer,
                 uint256 hashMul,
                 uint256 hashMod,
                 bytes memory fills,
                 uint256 amountOutMin
-            ) = abi.decode(data, (address, IERC20, uint256, bool, uint256, uint256, bytes, uint256));
+            ) = abi.decode(data, (address, IERC20, uint256, uint8, bool, uint256, uint256, bytes, uint256));
 
-            sellToUniswapV4(recipient, sellToken, ppm, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
+            sellToUniswapV4(recipient, sellToken, ppm, forkId, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
         } else if (action == uint32(ISettlerActions.BALANCERV3.selector)) {
             (
                 address recipient,
@@ -210,8 +211,12 @@ abstract contract ArbitrumMixin is
         return 0xbC0797015fcFc47d9C1856639CaE50D0e69FbEE8;
     }
 
-    function _POOL_MANAGER() internal pure override returns (IPoolManager) {
-        return ARBITRUM_POOL_MANAGER;
+    function _uniV4ForkInfo(uint8 forkId) internal pure override returns (IPoolManager poolManager) {
+        if (forkId == uniswapV4ForkId) {
+            poolManager = uniswapV4ArbitrumPoolManager;
+        } else {
+            revertUnknownForkId(forkId);
+        }
     }
 
     // I hate Solidity inheritance