+++ src/chains/Mainnet/TakerSubmitted.sol
@@ -56,8 +56,24 @@ contract MainnetSettler is Settler, MainnetMixin {
     function _dispatchVIP(uint256 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
         if (super._dispatchVIP(action, data)) {
             return true;
-        } else if ((action == uint32(ISettlerActions.UNISWAPV4_VIP.selector))
-                .or(action == uint32(ISettlerActions.BALANCERV3_VIP.selector))
+        } else if (action == uint32(ISettlerActions.UNISWAPV4_VIP.selector)) {
+            (
+                address recipient,
+                ISignatureTransfer.PermitTransferFrom memory permit,
+                uint8 forkId,
+                bool feeOnTransfer,
+                uint256 hashMul,
+                uint256 hashMod,
+                bytes memory fills,
+                bytes memory sig,
+                uint256 amountOutMin
+            ) = abi.decode(
+                data,
+                (address, ISignatureTransfer.PermitTransferFrom, uint8, bool, uint256, uint256, bytes, bytes, uint256)
+            );
+
+            sellToUniswapV4VIP(recipient, forkId, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
+        } else if ((action == uint32(ISettlerActions.BALANCERV3_VIP.selector))
                 .or(action == uint32(ISettlerActions.EKUBOV3_VIP.selector))) {
             (
                 address recipient,
@@ -72,9 +88,7 @@ contract MainnetSettler is Settler, MainnetMixin {
                 data, (address, ISignatureTransfer.PermitTransferFrom, bool, uint256, uint256, bytes, bytes, uint256)
             );
 
-            if (action == uint32(ISettlerActions.UNISWAPV4_VIP.selector)) {
-                sellToUniswapV4VIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
-            } else if (action == uint32(ISettlerActions.BALANCERV3_VIP.selector)) {
+            if (action == uint32(ISettlerActions.BALANCERV3_VIP.selector)) {
                 sellToBalancerV3VIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);
             } else { // if (action == uint32(ISettlerActions.EKUBOV3_VIP.selector))
                 sellToEkuboV3VIP(recipient, feeOnTransfer, hashMul, hashMod, fills, permit, sig, amountOutMin);