+++ src/core/univ4forks/Tsunami.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.25;
+
+import {IPoolManager} from "../UniswapV4Types.sol";
+
+IPoolManager constant tsunamiPoolManager = IPoolManager(0x6E4723A612831AfB9f5B2a5aE22723c37aAB9560);
+uint8 constant tsunamiForkId = 1;